app-service-posts/src/main/kotlin/com/github/arhor/aws/graphql/federation/posts/api/graphql/datafetcher/FederatedEntityFetcher.kt
package com.github.arhor.aws.graphql.federation.posts.api.graphql.datafetcher

import com.github.arhor.aws.graphql.federation.common.getUuid
import com.github.arhor.aws.graphql.federation.posts.generated.graphql.DgsConstants.USER
import com.github.arhor.aws.graphql.federation.posts.generated.graphql.types.User
import com.github.arhor.aws.graphql.federation.posts.service.UserRepresentationService
import com.netflix.graphql.dgs.DgsComponent
import com.netflix.graphql.dgs.DgsEntityFetcher

@DgsComponent
class FederatedEntityFetcher(
    private val userRepresentationService: UserRepresentationService,
) {

    /* Entity Fetchers */

    @DgsEntityFetcher(name = USER.TYPE_NAME)
    fun resolveUser(values: Map<String, Any>): User =
        userRepresentationService.findUserRepresentation(userId = values.getUuid(USER.Id))
}
